Fix various parser issues

+++ src/WebSocketImpl.cpp
@@ -15,7 +15,7 @@ bool WebSocketProtocol<isServer>::setCompressed(void *user) {
 }
 
 template <const bool isServer>
-bool WebSocketProtocol<isServer>::refusePayloadLength(void *user, int length) {
+bool WebSocketProtocol<isServer>::refusePayloadLength(void *user, uint64_t length) {
     return length > 16777216;
 }
